Added property on FDRequestClientTask to tell it wether or not the completion block should be automatically called on the main thread.
The completion blocks were always being called on whatever thread the
NSURLSession was using. In 99% of cases the user wants the completion
block to happen on the main thread so they can update the UI with new
data.

If the user wants the completion blocks to happen on a background
thread they can set this property to NO.