Removed FDRequestClient delegate and cache protocols.
These protocols are no longer necessary. The request client has block
parameters for all the functionality that the delegate used to provide
and if caching is needed on a per request basis we will add this
functionality directly to the request client. Apple does on disk
caching of URL requests since iOS 5 so I think for the immediately
future their caching and servers returning back correct cache headers
will work fine.