Created FDRequestClientTaskGroup class.
The FDRequestClientTaskGroup is analogous to dispatch groups. It allows
the user to group up multiple request client tasks and be notified once
all the tasks have been completed.

The grouped tasks are not dependent on one another. This is simply an
easy way to be notified when a group of tasks finish. Common use cases
are when multiple requests need to be made but the UI should only be
updated once all the requests have completed regardless of the order
they finished in.